Add Additional Podspec for Dynamic Framework
• We do not want to immediately switch all users to use our new fat dynamic framework
• However we do want it to be available on Cocoapods, so we've created an additional podspec & framework path
• When the project was started years back, the OneSignal-Dynamic target was actually creating a static framework, so I've renamed this target to OneSignal-Static-Framework
• The new OneSignal-Dynamic-Framework is just an aggregate framework. It does not directly produce its own product. Instead it uses a build script to produce a dynamic framework at repo/iOS_SDK/OneSignalSDK/Framework/Dynamic/OneSignal.framework
