+++ iOS_SDK/OneSignalSDK/Framework/OneSignal.framework/Headers
@@ -1,0 +1,1 @@
+Versions/Current/Headers
