+++ iOS_SDK/OneSignalSDK/Framework/OneSignal.framework/Modules
@@ -1,0 +1,1 @@
+Versions/Current/Modules
