--- iOS_SDK/OneSignalSDK/Framework/OneSignal.framework/Modules/module.modulemap
@@ -1,6 +1,0 @@
-framework module OneSignal {
-  umbrella header "OneSignal.h"
-
-  export *
-  module * { export * }
-}
