+++ iOS_SDK/OneSignalSDK/Framework/OneSignal.framework/Versions/A/Modules/module.modulemap
@@ -1,0 +1,5 @@
+framework module OneSignal {
+    umbrella header "OneSignal.h"
+    export *
+    module * { export * }
+}
