+++ iOS_SDK/OneSignalSDK/Framework/OneSignal.framework/Versions/Current
@@ -1,0 +1,1 @@
+A
